requirements_test.txt
# linters such as flake8 and pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version

aioresponses
codecov
coverage
mypy
pytest-cov
pytest-homeassistant-custom-component>=0.4.8
